+++ crates/ukis_h3cellstore/src/clickhouse/compacted_tables/schema/agg.rs
@@ -4,6 +4,7 @@ use serde::{Deserialize, Serialize};
 use crate::clickhouse::compacted_tables::schema::ClickhouseDataType;
 use crate::Named;
 
+/// Aggregations are only applied during aggregation, not compaction
 #[derive(Debug, PartialEq, Eq, Clone)]
 #[cfg_attr(feature = "use_serde", derive(Serialize, Deserialize))]
 pub enum AggregationMethod {
@@ -12,17 +13,19 @@ pub enum AggregationMethod {
     Max,
     Min,
     Average,
+    SetNull,
     // TODO: aggregation method to generate parent resolution for other h3index column
 }
 
 impl AggregationMethod {
-    pub fn is_applicable_to_datatype(&self, datatype: &ClickhouseDataType) -> bool {
+    pub fn is_applicable_to_datatype(&self, datatype: &ClickhouseDataType, nullable: bool) -> bool {
         match self {
             Self::RelativeToCellArea => !datatype.is_temporal() && datatype.is_number(),
             Self::Sum => !datatype.is_temporal() && datatype.is_number(),
             Self::Max => datatype.is_number(),
             Self::Min => datatype.is_number(),
             Self::Average => datatype.is_number(),
+            Self::SetNull => nullable,
         }
     }
 }
@@ -35,6 +38,7 @@ impl Named for AggregationMethod {
             Self::Min => "min",
             Self::Sum => "sum",
             Self::Average => "average",
+            Self::SetNull => "setnull",
         }
     }
 }