+++ crates/ukis_h3cellstorepy/tests/clickhouse/__init__.py
@@ -13,6 +13,7 @@ def elephant_schema(tableset_name="okavango_delta", temporal_partitioning="month
     csb.h3_partitioning(h3_partitioning, **kw)
     csb.add_column("is_valid", "UInt8", compression_method=CompressionMethod("gorilla"))
     csb.add_aggregated_column("elephant_density", "Float32", "RelativeToCellArea")
+    csb.add_aggregated_column("some_category", "UInt8", "SetNull", nullable=True)
     schema = csb.build()  # raises when the schema is invalid / missing something
     assert schema is not None
     #print(schema.to_json_string())
@@ -44,7 +45,8 @@ def setup_elephant_schema_with_data(clickhouse_grpc_endpoint, clickhouse_testdb_
     df = pl.DataFrame({
         "h3index": disk,
         "is_valid": np.ones(len(disk)),
-        "elephant_density": np.ones(len(disk)) * 4
+        "elephant_density": np.ones(len(disk)) * 4,
+        "some_category": np.ones(len(disk)) * 23
     })
 
     # write to db - this performs auto-compaction